Convert benchmarks to use cargo bench
Fix issues
