+++ src/examples/mod.rs
@@ -1,3 +1,2 @@
-pub mod benchmarks;
 pub mod test_rectangle_axis_aligned;
 pub mod test_rectangle_rotated;
